changed view url to solutions/id from solutions/id/published
